refactor: `firebase_options` to new utils directory
